We now use DIP coordinates on all ports

+++ boden/include/bdn/Point.h
@@ -11,14 +11,14 @@ namespace bdn
 struct Point
 {
 public:
-	int x = 0;
-	int y = 0;
+	double x = 0;
+	double y = 0;
 
 	Point()
 	{		
 	}
 
-	Point(int x, int y)
+	Point(double x, double y)
 		: x(x)
 		, y(y)
     {
